refactor(tdigest): replace template Rank with separate Rank and RevRank methods
