+++ tests/cppunit/types/tdigest_test.cc
@@ -312,7 +312,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_the_set_containing_different_elemen
   std::vector<int> result;
   result.reserve(input.size());
   const std::vector<double> value = {0, 10, 20, 30, 40, 50, 60, 70};
-  status = tdigest_->Rank<true>(*ctx_, test_digest_name, value, result);
+  status = tdigest_->RevRank(*ctx_, test_digest_name, value, result);
   const auto expect_result_revrank = std::vector<double>{6, 5, 4, 3, 2, 1, 0, -1};
 
   for (size_t i = 0; i < result.size(); i++) {
@@ -323,7 +323,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_the_set_containing_different_elemen
 
   result.clear();
   result.reserve(input.size());
-  status = tdigest_->Rank<false>(*ctx_, test_digest_name, value, result);
+  status = tdigest_->Rank(*ctx_, test_digest_name, value, result);
   const auto expect_result_rank = std::vector<double>{-1, 0, 1, 2, 3, 4, 5, 6};
   for (size_t i = 0; i < result.size(); i++) {
     auto got = result[i];
@@ -345,7 +345,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_the_set_containing_several_identica
   std::vector<int> result;
   const std::vector<double> value = {10, 20};
   result.reserve(value.size());
-  status = tdigest_->Rank<true>(*ctx_, test_digest_name, value, result);
+  status = tdigest_->RevRank(*ctx_, test_digest_name, value, result);
   const auto expect_result_revrank = std::vector<double>{3, 1};
   for (size_t i = 0; i < result.size(); i++) {
     auto got = result[i];
@@ -355,7 +355,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_the_set_containing_several_identica
 
   result.clear();
   result.reserve(value.size());
-  status = tdigest_->Rank<false>(*ctx_, test_digest_name, value, result);
+  status = tdigest_->Rank(*ctx_, test_digest_name, value, result);
   const auto expect_result_rank = std::vector<double>{1, 4};
   for (size_t i = 0; i < result.size(); i++) {
     auto got = result[i];
@@ -368,7 +368,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_the_set_containing_several_identica
 
   result.clear();
   result.reserve(value.size());
-  status = tdigest_->Rank<true>(*ctx_, test_digest_name, value, result);
+  status = tdigest_->RevRank(*ctx_, test_digest_name, value, result);
   const auto expect_result_new_revrank = std::vector<double>{4, 1};
   for (size_t i = 0; i < result.size(); i++) {
     auto got = result[i];
@@ -378,7 +378,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_the_set_containing_several_identica
 
   result.clear();
   result.reserve(value.size());
-  status = tdigest_->Rank<false>(*ctx_, test_digest_name, value, result);
+  status = tdigest_->Rank(*ctx_, test_digest_name, value, result);
   const auto expect_result_new_rank = std::vector<double>{2, 5};
   for (size_t i = 0; i < result.size(); i++) {
     auto got = result[i];
@@ -397,7 +397,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_empty_tdigest) {
   std::vector<int> result;
   result.reserve(2);
   const std::vector<double> value = {10, 20};
-  status = tdigest_->Rank<true>(*ctx_, test_digest_name, value, result);
+  status = tdigest_->RevRank(*ctx_, test_digest_name, value, result);
   const auto expect_result_revrank = std::vector<double>{-2, -2};
   for (size_t i = 0; i < result.size(); i++) {
     auto got = result[i];
@@ -407,7 +407,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_empty_tdigest) {
 
   result.clear();
   result.reserve(2);
-  status = tdigest_->Rank<false>(*ctx_, test_digest_name, value, result);
+  status = tdigest_->Rank(*ctx_, test_digest_name, value, result);
   const auto expect_result_rank = std::vector<double>{-2, -2};
   for (size_t i = 0; i < result.size(); i++) {
     auto got = result[i];
@@ -430,7 +430,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_different_or_same_and_unordered_inp
   std::vector<int> result;
   const std::vector<double> value = {50, 36, 4, 99, 8.8};
   result.reserve(value.size());
-  status = tdigest_->Rank<false>(*ctx_, test_digest_name, value, result);
+  status = tdigest_->Rank(*ctx_, test_digest_name, value, result);
   const auto expect_result_rank = std::vector<double>{13, 11, 1, 16, 4};
   for (size_t i = 0; i < result.size(); i++) {
     auto got = result[i];
@@ -441,7 +441,7 @@ TEST_F(RedisTDigestTest, RevRank_and_Rank_on_different_or_same_and_unordered_inp
   const std::vector<double> value_new = {50, 36, 4, 99, 8.8, 12};
   result.clear();
   result.reserve(value_new.size());
-  status = tdigest_->Rank<true>(*ctx_, test_digest_name, value_new, result);
+  status = tdigest_->RevRank(*ctx_, test_digest_name, value_new, result);
   const auto expect_result_revrank = std::vector<double>{4, 7, 16, 1, 14, 12};
   for (size_t i = 0; i < result.size(); i++) {
     auto got = result[i];
